+++ packages/core/src/columns/code.ts
@@ -1,5 +1,6 @@
-import { toArray } from '../utils';
+import { singleQuote, toArray } from '../utils';
 import { ColumnChain } from './columnType';
+import { isRaw, RawExpression } from '../raw';
 
 export type Code = string | Code[];
 
@@ -40,4 +41,45 @@ export const columnChainToCode = (
   return result.length === 1 && typeof result[0] === 'string'
     ? result[0]
     : result;
+};
+
+export const codeToString = (
+  code: Code,
+  tabs: string,
+  shift: string,
+): string => {
+  if (typeof code === 'string') return `${tabs}${code}`;
+
+  const lines: string[] = [];
+  for (const item of code) {
+    if (typeof item === 'string') {
+      lines.push(`${tabs}${item}`);
+    } else {
+      lines.push(codeToString(item, tabs + shift, shift));
+    }
+  }
+
+  return lines.length ? lines.join('\n') : '';
+};
+
+export const columnDefaultArgumentToCode = (
+  t: string,
+  value: unknown,
+): string => {
+  if (typeof value === 'object' && value && isRaw(value)) {
+    return rawToCode(t, value);
+  } else if (typeof value === 'function') {
+    return value.toString();
+  } else if (typeof value === 'string') {
+    return singleQuote(value);
+  } else {
+    return JSON.stringify(value);
+  }
+};
+
+export const rawToCode = (t: string, raw: RawExpression): string => {
+  const values = raw.__values;
+  return `${t}.raw(${singleQuote(raw.__raw)}${
+    values ? `, ${JSON.stringify(values)}` : ''
+  })`;
 };