+++ packages/docs/src/.vitepress/config.ts
@@ -2,62 +2,66 @@ export default {
   title: 'Orchid ORM',
   description: 'Postgres ORM & Query Builder',
   head: [
-    ["script", { async: true, src: 'https://www.googletagmanager.com/gtag/js?id=G-PV4PL9TK79' }],
+    [
+      'script',
+      {
+        async: true,
+        src: 'https://www.googletagmanager.com/gtag/js?id=G-PV4PL9TK79',
+      },
+    ],
   ],
   markdown: {
     theme: 'one-dark-pro',
   },
   themeConfig: {
-    nav: [
-      { text: 'Guide', link: '/guide/', activeMatch: '^/guide/' },
-    ],
+    nav: [{ text: 'Guide', link: '/guide/', activeMatch: '^/guide/' }],
     sidebar: [
       {
         items: [
           {
             text: 'Overview',
-            link: '/guide/'
+            link: '/guide/',
           },
           {
             text: 'Quickstart',
-            link: '/guide/quickstart'
+            link: '/guide/quickstart',
           },
           {
             text: 'Building a sample app',
-            link: '/guide/building-a-sample-app'
+            link: '/guide/building-a-sample-app',
           },
           {
             text: 'Benchmarks',
-            link: '/guide/benchmarks'
+            link: '/guide/benchmarks',
           },
           {
             text: 'Current status and limitations',
-            link: '/guide/current-status-and-limitations'
-          }
-        ]
+            link: '/guide/current-status-and-limitations',
+          },
+        ],
       },
       {
         text: 'Query builder',
         items: [
           {
             text: 'Setup and overview',
-            link: '/guide/query-builder-setup'
+            link: '/guide/query-builder-setup',
           },
           {
             text: 'Query methods',
-            link: '/guide/query-builder'
+            link: '/guide/query-builder',
           },
           {
             text: 'Where conditions',
-            link: '/guide/query-builder-where'
+            link: '/guide/query-builder-where',
           },
           {
-            text: 'Create, Update, Delete records',
-            link: '/guide/query-builder-create-update-delete'
+            text: 'Create, Update, Delete',
+            link: '/guide/query-builder-create-update-delete',
           },
           {
             text: 'Aggregate functions',
-            link: '/guide/query-builder-aggregate'
+            link: '/guide/query-builder-aggregate',
           },
           {
             text: 'Advanced methods',
@@ -65,38 +69,38 @@ export default {
           },
           {
             text: 'Callbacks',
-            link: '/guide/query-builder-callbacks'
+            link: '/guide/query-builder-callbacks',
           },
           {
             text: 'Error handling',
-            link: '/guide/query-builder-error-handling'
+            link: '/guide/query-builder-error-handling',
           },
-        ]
+        ],
       },
       {
         text: 'ORM',
         items: [
           {
             text: 'Setup and overview',
-            link: '/guide/orm-setup-and-overview'
+            link: '/guide/orm-setup-and-overview',
           },
           {
             text: 'Modeling relations',
-            link: '/guide/orm-relations'
+            link: '/guide/orm-relations',
           },
           {
             text: 'Relation queries',
-            link: '/guide/orm-relation-queries'
+            link: '/guide/orm-relation-queries',
           },
           {
             text: 'Repository',
-            link: '/guide/orm-repo'
+            link: '/guide/orm-repo',
           },
           {
             text: 'Test factories',
-            link: '/guide/orm-test-factories'
-          }
-        ]
+            link: '/guide/orm-test-factories',
+          },
+        ],
       },
       {
         text: 'Columns schema',
@@ -121,7 +125,7 @@ export default {
             text: 'JSON types',
             link: '/guide/columns-json-types',
           },
-        ]
+        ],
       },
       {
         text: 'Migrations',
@@ -142,8 +146,8 @@ export default {
             text: 'Writing a migration',
             link: '/guide/migration-writing',
           },
-        ]
-      }
+        ],
+      },
     ],
   },
-}
+};