+++ packages/orm/src/relations/hasMany.ts
@@ -214,7 +214,7 @@ export const makeHasManyMethod = (
       return (query) => {
         const fromQuery = query.clone();
         fromQuery.query.select = fromQuerySelect;
-        (relationQuery.query as InsertQueryData).fromQuery = fromQuery;
+        (relationQuery.query as InsertQueryData).values = { from: fromQuery };
       };
     },
   };